+++ config/webpack.script.config.js
@@ -1,6 +1,8 @@
+const path = require('path')
+
 module.exports = [
   {
-    test: /\.tsx$/,
+    test: [/\.jsx$/, /\.tsx$/],
     exclude: /node_modules/,
     use: [
       'thread-loader',
@@ -16,25 +18,18 @@ module.exports = [
               }
             ],
             '@babel/preset-typescript'
-          ]
-        }
-      }
-    ]
-  },
-  {
-    test: /\.jsx$/,
-    exclude: /node_modules/,
-    use: [
-      'thread-loader',
-      {
-        loader: 'babel-loader',
-        options: {
-          cacheDirectory: true,
-          presets: [
+          ],
+          plugins: [
             [
-              '@babel/preset-react',
+              'react-css-modules',
               {
-                runtime: 'automatic'
+                context: path.join(__dirname, '..'),
+                exclude: 'node_modules',
+                filetypes: {
+                  '.less': {
+                    syntax: 'postcss-less'
+                  }
+                }
               }
             ]
           ]