+++ config/webpack.styles.config.js
@@ -1,6 +1,6 @@
 module.exports = [
   {
-    test: /\.less$/,
+    test: [/\.less$/, /\.css$/],
     use: [
       'thread-loader',
       'style-loader',
@@ -11,28 +11,6 @@ module.exports = [
         }
       },
       'less-loader',
-      {
-        loader: 'postcss-loader',
-        options: {
-          postcssOptions: {
-            // 添加 autoprefixer 插件
-            plugins: [require('autoprefixer')]
-          }
-        }
-      }
-    ]
-  },
-  {
-    test: /\.css$/,
-    use: [
-      'thread-loader',
-      'style-loader',
-      {
-        loader: 'css-loader',
-        options: {
-          importLoaders: 1
-        }
-      },
       {
         loader: 'postcss-loader',
         options: {